pdf-extractor-python/requirements.txt
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0

